fix(*): refactor code, add ParserCore
