+++ example.txt
@@ -1,1 +1,1 @@
-"dwdwdwdwfwfrbtjlyiuoi"
+"AAA"
